requirements.txt
Django
django-filer
django-mptt
easy_thumbnails
django-polymorphic
